# api/openapi.yaml
openapi: 3.0.3
info:
  title: Journey Platform API
  description: Event-Driven Journey Engine & Workflow Platform REST API Specification
  version: 1.0.0
servers:
  - url: /api/v1
    description: Base v1 API path
paths:
...
  /journeys/drafts:
    post:
      summary: Create a new journey draft
      operationId: createJourneyDraft
      tags:
        - Journey Drafts
      parameters:
        - $ref: '#/components/parameters/HeaderRequestID'
        - $ref: '#/components/parameters/HeaderIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GraphDraft'
      responses:
        '201':
          description: Journey draft created successfully
          headers:
            ETag:
              $ref: '#/components/headers/HeaderETag'
            Request-ID:
              $ref: '#/components/headers/HeaderRequestIDResponse'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphDraft'
        '400':
          $ref: '#/components/responses/ErrorResponse400'
        '500':
          $ref: '#/components/responses/ErrorResponse500'
  /journeys/drafts/{draft_id}:
    get:
      summary: Get journey draft details
      operationId: getJourneyDraft
      tags:
        - Journey Drafts
      parameters:
        - name: draft_id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/HeaderRequestID'
      responses:
        '200':
          description: Journey draft details
          headers:
            ETag:
              $ref: '#/components/headers/HeaderETag'
            Request-ID:
              $ref: '#/components/headers/HeaderRequestIDResponse'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphDraft'
        '404':
          $ref: '#/components/responses/ErrorResponse404'
        '500':
          $ref: '#/components/responses/ErrorResponse500'

    put:
      summary: Edit/update journey draft
      operationId: updateJourneyDraft
      tags:
        - Journey Drafts
      parameters:
        - name: draft_id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/HeaderRequestID'
        - $ref: '#/components/parameters/HeaderIfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GraphDraft'
      responses:
        '200':
          description: Journey draft updated
          headers:
            ETag:
              $ref: '#/components/headers/HeaderETag'
            Request-ID:
              $ref: '#/components/headers/HeaderRequestIDResponse'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphDraft'
        '400':
          $ref: '#/components/responses/ErrorResponse400'
        '412':
          $ref: '#/components/responses/ErrorResponse412'
        '500':
          $ref: '#/components/responses/ErrorResponse500'
...
components:
  parameters:
    HeaderRequestID:
      name: Request-ID
      in: header
      required: false
      description: Unique request identifier for distributed tracing
      schema:
        type: string
        example: "req-12345678"
    HeaderIdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Key for safe retry of mutating requests
      schema:
        type: string
        example: "idem-87654321"
    HeaderIfMatch:
      name: If-Match
      in: header
      required: false
      description: ETag value for optimistic concurrency check
      schema:
        type: string
        example: '"1.0.0-hash"'
  headers:
    HeaderRequestIDResponse:
      description: Unique request identifier returned by the server
      schema:
        type: string
        example: "req-12345678"
    HeaderETag:
      description: Resource state fingerprint for concurrency and caching
      schema:
        type: string
        example: '"1.0.0-hash"'
  responses:
    ErrorResponse400:
      description: Invalid client request body or parameters
      headers:
        Request-ID:
          $ref: '#/components/headers/HeaderRequestIDResponse'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ErrorResponse404:
      description: Requested resource not found
      headers:
        Request-ID:
          $ref: '#/components/headers/HeaderRequestIDResponse'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ErrorResponse412:
      description: Precondition failed (ETag / If-Match mismatch)
      headers:
        Request-ID:
          $ref: '#/components/headers/HeaderRequestIDResponse'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ErrorResponse500:
      description: Internal server error
      headers:
        Request-ID:
          $ref: '#/components/headers/HeaderRequestIDResponse'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    GraphDraft:
      type: object
      required:
        - schema_version
        - draft_id
        - tenant_id
        - name
        - version
        - nodes
        - edges
      properties:
        schema_version:
          type: string
          enum: ["1.0"]
        draft_id:
          type: string
        tenant_id:
          type: string
        name:
          type: string
        description:
          type: string
        version:
          type: integer
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/GraphNode'
        edges:
          type: array
          items:
            $ref: '#/components/schemas/GraphEdge'
        content_hash:
          type: string
        created_at:
          type: string
        updated_at:
          type: string
      additionalProperties: false
    GraphNode:
      type: object
      required:
        - id
        - type
        - name
      properties:
        id:
          type: string
        type:
          type: string
        name:
          type: string
        config:
          type: object
        position:
          type: object
          properties:
            x:
              type: number
            y:
              type: number
      additionalProperties: false
    GraphEdge:
      type: object
      required:
        - id
        - source
        - target
      properties:
        id:
          type: string
        source:
          type: string
        target:
          type: string
        condition:
          type: string
      additionalProperties: false
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
        - request_id
      properties:
        code:
          type: string
          example: "INVALID_REQUEST"
        message:
          type: string
          example: "Validation failed"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              issue:
                type: string
        request_id:
          type: string
          example: "req-12345678"
      additionalProperties: false